Refactor mock classes
