+++ sdks/cpp/common/tests/.gitignore
@@ -1,4 +1,5 @@
 *
 !*.cpp
+!*.h
 !CMakeLists.txt
 !.gitignore
